backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic[email]
python-multipart==0.0.9
numpy>=1.23
pandas>=2.0
optuna==4.7.0
scikit-learn
sqlalchemy>=2.0
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-dotenv
alembic
eval_type_backport
asyncpg
authlib
httpx
itsdangerous
yfinance
# ── Sentiment Intelligence Layer ───────────────────────────────────────────────
vaderSentiment
transformers>=4.38
torch --index-url https://download.pytorch.org/whl/cpu
newsapi-python
beautifulsoup4
python-dotenv
